Changed the Create New Workflow component to be inline
